Update Readme.md to say what this isn't
I'm not (right now?) interested in making anything that could be considered a full-fledged typing practice program.  This is simply something that I'm using to force myself to read by typing out whatever text it is I'm trying to read.  A lot of typing programs that enable you to supply custom text get hung up on special characters and stress me out about mistakes and speed, which for me is not the point.